% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bs_block_names}
\alias{bs_block_names}
\title{Block-level parameter names (no transformed parameters / generated
quantities), dot-indexed. Length equals \code{bs_ndim_block()}.}
\usage{
bs_block_names(handle)
}
\description{
Block-level parameter names (no transformed parameters / generated
quantities), dot-indexed. Length equals \code{bs_ndim_block()}.
}
\keyword{internal}
