% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bs_full_names}
\alias{bs_full_names}
\title{Full constrained parameter names (block + transformed parameters +
generated quantities), dot-indexed.}
\usage{
bs_full_names(handle)
}
\description{
Full constrained parameter names (block + transformed parameters +
generated quantities), dot-indexed.
}
\keyword{internal}
