% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bs_ndim_block}
\alias{bs_ndim_block}
\title{Number of block-level constrained parameters (no TP, no GQ).}
\usage{
bs_ndim_block(handle)
}
\description{
Number of block-level constrained parameters (no TP, no GQ).
}
\keyword{internal}
