% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bs_open}
\alias{bs_open}
\title{Open a BridgeStan model and return an \verb{ExternalPtr<BSHandle>} that caches
parameter-name metadata. The handle may be used by any of the \verb{bs_*}
accessor functions without re-opening the shared library.}
\usage{
bs_open(lib_path, data_json, seed)
}
\description{
Open a BridgeStan model and return an \verb{ExternalPtr<BSHandle>} that caches
parameter-name metadata. The handle may be used by any of the \verb{bs_*}
accessor functions without re-opening the shared library.
}
\keyword{internal}
