% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bs_param_constrain}
\alias{bs_param_constrain}
\title{Map an unconstrained position to the full constrained scale (including
transformed parameters and generated quantities) using an already-opened
handle.}
\usage{
bs_param_constrain(handle, theta_unc, seed)
}
\description{
Map an unconstrained position to the full constrained scale (including
transformed parameters and generated quantities) using an already-opened
handle.
}
\keyword{internal}
