% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bs_param_unconstrain_json}
\alias{bs_param_unconstrain_json}
\title{Map a constrained point (JSON, CmdStan format) to the unconstrained space,
using an already-opened handle. All parameters must be present in the JSON.}
\usage{
bs_param_unconstrain_json(handle, init_json)
}
\description{
Map a constrained point (JSON, CmdStan format) to the unconstrained space,
using an already-opened handle. All parameters must be present in the JSON.
}
\keyword{internal}
