% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bs_unc_names}
\alias{bs_unc_names}
\title{Unconstrained parameter names, dot-indexed. Length equals \code{bs_ndim_unc()}.}
\usage{
bs_unc_names(handle)
}
\description{
Unconstrained parameter names, dot-indexed. Length equals \code{bs_ndim_unc()}.
}
\keyword{internal}
