% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{nutpie_warmup_diagnostics}
\alias{nutpie_warmup_diagnostics}
\title{Extract warmup diagnostics from nutpie output}
\usage{
nutpie_warmup_diagnostics(draws)
}
\arguments{
\item{draws}{A \code{posterior::draws_array} returned by \code{\link[=nutpie_sample]{nutpie_sample()}} with
\code{save_warmup = TRUE}.}
}
\value{
A named list of diagnostic vectors for the warmup phase.
}
\description{
Extract warmup diagnostics from nutpie output
}
