% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{nutpie_warmup_draws}
\alias{nutpie_warmup_draws}
\title{Extract warmup draws from nutpie output}
\usage{
nutpie_warmup_draws(draws)
}
\arguments{
\item{draws}{A \code{posterior::draws_array} returned by \code{\link[=nutpie_sample]{nutpie_sample()}} with
\code{save_warmup = TRUE}.}
}
\value{
A \code{posterior::draws_array} containing the warmup draws, or \code{NULL}
if warmup draws were not saved.
}
\description{
Extract warmup draws from nutpie output
}
