% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutpieR-package.R
\docType{package}
\name{nutpieR-package}
\alias{nutpieR}
\alias{nutpieR-package}
\title{nutpieR: R Bindings for the Nutpie NUTS Sampler}
\description{
Lightweight R interface to the nutpie No-U-Turn Sampler (NUTS) implemented in Rust (nuts-rs), using BridgeStan for model gradients.
}
\author{
\strong{Maintainer}: Tim Mastny \email{tim.mastny@gmail.com}

}
\keyword{internal}
