% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{param_constrain_rs}
\alias{param_constrain_rs}
\title{Map an unconstrained position to the constrained scale (including
transformed parameters and generated quantities).}
\usage{
param_constrain_rs(lib_path, data_json, theta_unc, seed)
}
\description{
Map an unconstrained position to the constrained scale (including
transformed parameters and generated quantities).
}
\keyword{internal}
