% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{param_unconstrain_json_rs}
\alias{param_unconstrain_json_rs}
\title{Map a constrained point (JSON, CmdStan format) to the unconstrained space.
All parameters must be present in the JSON.}
\usage{
param_unconstrain_json_rs(lib_path, data_json, init_json)
}
\description{
Map a constrained point (JSON, CmdStan format) to the unconstrained space.
All parameters must be present in the JSON.
}
\keyword{internal}
