% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sample_normal}
\alias{sample_normal}
\title{Sample from a 10-d standard normal using nuts-rs.
Returns a matrix of draws (rows = draws*chains, cols = parameters).}
\usage{
sample_normal(num_draws, num_chains, seed)
}
\arguments{
\item{num_draws}{Number of draws per chain after warmup.}

\item{num_chains}{Number of parallel chains.}

\item{seed}{Random seed.}
}
\description{
Sample from a 10-d standard normal using nuts-rs.
Returns a matrix of draws (rows = draws*chains, cols = parameters).
}
\keyword{internal}
