% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nutpie_clear_cache}
\alias{nutpie_clear_cache}
\title{Clear the nutpieR compile cache}
\usage{
nutpie_clear_cache()
}
\value{
Invisibly \code{NULL}.
}
\description{
Removes the entire compile cache tree under
\code{\link[=nutpie_cache_dir]{nutpie_cache_dir()}}. All cached compiled models
will be recompiled on next use.
}
\examples{
nutpie_clear_cache()
}
