% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{nutpie_compile_model}
\alias{nutpie_compile_model}
\title{Compile a Stan model}
\usage{
nutpie_compile_model(stan_file)
}
\arguments{
\item{stan_file}{Path to a \code{.stan} file.}
}
\value{
An object of class \code{"nutpie_model"} containing the path to the
compiled shared library.
}
\description{
Compiles a Stan model to a shared library using BridgeStan.
Downloads BridgeStan sources on first use (this is slow).
}
